refactor directories
all items go into directories specified in notes. direnv and venv support added.
