+++ .envrc
@@ -1,0 +1,2 @@
+export VIRTUAL_ENV=".venv"
+layout python python3
