+++ .gitignore
@@ -1,0 +1,4 @@
+data_raw/*
+data_clean/*.pt
+.venv
+src/__pycache__/*
