+++ requirements.txt
@@ -1,0 +1,46 @@
+aiohappyeyeballs==2.4.3
+aiohttp==3.10.9
+aiosignal==1.3.1
+attrs==24.2.0
+certifi==2024.8.30
+charset-normalizer==3.3.2
+filelock==3.16.1
+frozenlist==1.4.1
+fsspec==2024.9.0
+idna==3.10
+imageio==2.35.1
+Jinja2==3.1.4
+lazy_loader==0.4
+MarkupSafe==2.1.5
+mpmath==1.3.0
+multidict==6.1.0
+networkx==3.3
+numpy==2.1.2
+nvidia-cublas-cu12==12.1.3.1
+nvidia-cuda-cupti-cu12==12.1.105
+nvidia-cuda-nvrtc-cu12==12.1.105
+nvidia-cuda-runtime-cu12==12.1.105
+nvidia-cudnn-cu12==9.1.0.70
+nvidia-cufft-cu12==11.0.2.54
+nvidia-curand-cu12==10.3.2.106
+nvidia-cusolver-cu12==11.4.5.107
+nvidia-cusparse-cu12==12.1.0.106
+nvidia-nccl-cu12==2.20.5
+nvidia-nvjitlink-cu12==12.6.77
+nvidia-nvtx-cu12==12.1.105
+packaging==24.1
+pillow==10.4.0
+psutil==6.0.0
+pyparsing==3.1.4
+requests==2.32.3
+scikit-image==0.24.0
+scipy==1.14.1
+sympy==1.13.3
+tifffile==2024.9.20
+torch==2.4.1
+torch-geometric==2.6.1
+tqdm==4.66.5
+triton==3.0.0
+typing_extensions==4.12.2
+urllib3==2.2.3
+yarl==1.13.1
